requirements.txt
libsass==0.23.0
setuptools~=75.3.0
pandas~=2.2.3
Jinja2~=3.1.2
pathvalidate~=3.2.0
weasyprint~=60.1
fuzzywuzzy~=0.18.0
pdfgen~=1.0.5
pyppeteer~=0.2.6
mailjet-rest==1.3.4
selenium~=4.26.1
wakepy~=0.7.2
pngquant~=1.0.7
pyyaml==6.0.1